Grunt and related config file organization
