+++ config/grunt/release.js
@@ -1,0 +1,28 @@
+/*global module */
+module.exports = function (grunt) {
+    'use strict';
+
+    grunt.config.merge({
+        bump: {
+            options: {
+                files: [
+                    'bower.json',
+                    'package.json'
+                ],
+                updateConfigs: [],
+                commit: true,
+                commitMessage: 'Release v%VERSION%',
+                commitFiles: ['-a'],
+                createTag: false,
+                tagName: 'v%VERSION%',
+                tagMessage: 'Version %VERSION%',
+                push: true,
+                pushTo: 'origin',
+                gitDescribeOptions: '--tags --always --abbrev=1 --dirty=-d',
+                globalReplace: false,
+                prereleaseName: false,
+                regExp: false
+            }
+        }
+    });
+};
