+++ config/karma/karma.conf-ci.js
