+++ config/wdio/wdio.conf-shared.js
